Revert "fix: stabilize homepage platform and GitHub stats (#4981)"
This reverts commit 09362b45e84d093a3fb167d9ccdc09ca1acd644b.
